Performance test (#75)
* Reference calculation taken from https://people.sc.fsu.edu/~jburkardt/f_src/timer/timer.html
in order to stablish a reference CPU time

* Changed time to user time

* Included performance expected times

* Included performance expected times only positive check

* Banner indicating the test will fail if performance goes down

* Updated travis with performance

* Updated travis with performance again

* Update travis again and again

+++ tests/run_test.sh
@@ -4,14 +4,60 @@
 
 MY_PATH=`pwd`                                   # Capturing the local path of the folder where we are running.
 
+declare -A performanceExpectedTimes
+performanceExpectedTimes["single.point"]=0.014
+performanceExpectedTimes["single.point.noelec"]=0.027
+performanceExpectedTimes["single.point.rspace"]=0.036
+performanceExpectedTimes["opt"]=0.127
+performanceExpectedTimes["opt.cg"]=0.129
+performanceExpectedTimes["opt_cons"]=0.127
+performanceExpectedTimes["dorbitals"]=0.412
+performanceExpectedTimes["tableread"]=1.700
+performanceExpectedTimes["0scf"]=0.127
+performanceExpectedTimes["2scf"]=0.120
+performanceExpectedTimes["fullscf"]=0.158
+performanceExpectedTimes["fullscf.etemp"]=0.160
+performanceExpectedTimes["sp2"]=0.161
+performanceExpectedTimes["sp2.sparse"]=0.579
+performanceExpectedTimes["fullscf.nvt"]=0.158
+performanceExpectedTimes["fullscf.npt"]=0.157
+performanceExpectedTimes["fullscf.vdw"]=0.158
+performanceExpectedTimes["fullscf.spin"]=0.231
+performanceExpectedTimes["fullscf.kon"]=0.973
+performanceExpectedTimes["fullscf.rspace"]=0.014
+performanceExpectedTimes["fittingoutput.dat"]=0.002
+performanceExpectedTimes["0scf"]=0.127
+performanceExpectedTimes["fullscf"]=0.158
+performanceExpectedTimes["sp2"]=0.161
+
 RUN="./LATTE_DOUBLE"  # LATTE program executable
 
-echo -e "\nTesting LATTE with new (latte.in) input files \n"
+echo ""
+echo "Reference elapsed CPU time"
+echo "=========================="
+echo ""
+
+( cd tests; make; cd .. ) &> /dev/null
+
+echo -en "   Testing for reference ... "
+timeRef=`( /usr/bin/time -f "%U" ./tests/timer_cpu_time > out ) 2>&1 > /dev/null`
+
+relativeTime=`echo "$timeRef/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+echo "(${timeRef}s,$relativeTime) PASSED"
+
+if [ "$LATTE_PERFORMANCE" = "yes" ]
+then
+	echo ">> If performance goes lower than 10% the test will fail"
+fi
+
+echo ""
+echo "Testing LATTE with new (latte.in) input files"
+echo "============================================="
+echo ""
 
 mv latte.in latte.in.tmp
 
 set -e                                          # This will exit the script if there is any error
-
 
 # Testing for single point calculations:
 
@@ -27,12 +73,19 @@ for name in single.point single.point.noelec single.point.rspace ; do
 
   echo -en "   Testing for "$name" ... "
 
-  time=`( /usr/bin/time -f "%S" $RUN > out ) 2>&1 > /dev/null`
+  time=`( /usr/bin/time -f "%U" $RUN > out ) 2>&1 > /dev/null`
   ENERG=`grep -e "FREE ENERGY" out | awk 'NF>1{print $5}'`
   echo $ENERG > energy.out
 
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-energy.py --reference $REF --current energy.out --reltol 0.00001
+  
+  if [ "$LATTE_PERFORMANCE" = "yes" ]
+  then
+	test=`echo "($relativeTime-${performanceExpectedTimes[$name]})/${performanceExpectedTimes[$name]} < 0.1" | bc -l`
+	[ "$test" -eq 0 ] && exit -1
+  fi
 
   rm $REF out
 
@@ -55,10 +108,17 @@ for name in opt opt.cg opt_cons dorbitals; do
 
   echo -en "   Testing for "$name" ... "
 
-  time=`( /usr/bin/time -f "%S" $RUN > out ) 2>&1 > /dev/null`
+  time=`( /usr/bin/time -f "%U" $RUN > out ) 2>&1 > /dev/null`
 
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-optim.py --reference $REF --current monitorrelax.xyz --reltol 0.00001
+  
+  if [ "$LATTE_PERFORMANCE" = "yes" ]
+  then
+	test=`echo "($relativeTime-${performanceExpectedTimes[$name]})/${performanceExpectedTimes[$name]} < 0.1" | bc -l`
+	[ "$test" -eq 0 ] && exit -1
+  fi
 
   #rm $REF monitorrelax.xyz out
 done
@@ -78,13 +138,20 @@ for name in tableread 0scf 2scf fullscf fullscf.etemp sp2 sp2.sparse fullscf.nvt
 
   echo -en "   Testing for "$name" ... "
 
-  time=`( /usr/bin/time -f "%S" $RUN > out ) 2>&1 > /dev/null`
+  time=`( /usr/bin/time -f "%U" $RUN > out ) 2>&1 > /dev/null`
   
   grep "Data" out | sed 's/Data/ /g' | awk 'NF>1{print $2}' > energy.out
 
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-energy.py --reference $REF --current energy.out --reltol 0.00001
 
+  if [ "$LATTE_PERFORMANCE" = "yes" ]
+  then
+	test=`echo "($relativeTime-${performanceExpectedTimes[$name]})/${performanceExpectedTimes[$name]} < 0.1" | bc -l`
+	[ "$test" -eq 0 ] && exit -1
+  fi
+  
   rm $REF energy.out out
 
 done
@@ -103,7 +170,7 @@ for name in fittingoutput.dat ; do
 
   echo -en "   Testing for "$name" ... "
 
-  time=`( /usr/bin/time -f "%S" $RUN > out ) 2>&1 > /dev/null`
+  time=`( /usr/bin/time -f "%U" $RUN > out ) 2>&1 > /dev/null`
   
   tol=0.0001
   
@@ -131,7 +198,8 @@ for name in fittingoutput.dat ; do
       }
     ' $REF`
   
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   if [ "$check" -eq 0 ]
   then
     echo "PASSED"
@@ -139,6 +207,12 @@ for name in fittingoutput.dat ; do
     echo ""
     diff $REF $name
     exit -1
+  fi
+  
+  if [ "$LATTE_PERFORMANCE" = "yes" ]
+  then
+	test=`echo "($relativeTime-${performanceExpectedTimes[$name]})/${performanceExpectedTimes[$name]} < 0.1" | bc -l`
+	[ "$test" -eq 0 ] && exit -1
   fi
 
   rm $REF out
@@ -149,7 +223,10 @@ done
 
 rm latte.in
 
-echo -e "\nTesting LATTE with original input files \n"
+echo ""
+echo "Testing LATTE with original input files"
+echo "======================================="
+echo ""
 
 for name in 0scf fullscf sp2 ; do
 
@@ -165,13 +242,20 @@ for name in 0scf fullscf sp2 ; do
 
   echo -en "   Testing for "$name" ... "
 
-  time=`( /usr/bin/time -f "%S" $RUN > out ) 2>&1 > /dev/null`
+  time=`( /usr/bin/time -f "%U" $RUN > out ) 2>&1 > /dev/null`
   
   grep "Data" out | sed 's/Data/ /g' | awk 'NF>1{print $2}' > energy.out
   grep Energy out | sed -e s/"PAR"/$STRR/g  >  input_tmp.in
   
-  echo -n "($time s) "
+  relativeTime=`echo "$time/$timeRef" | bc -l | awk '{printf("%.3f",$1)}'`
+  echo -n "(${time}s,$relativeTime) "
   python ./tests/test-energy.py --reference $REF --current energy.out --reltol 0.00001
+  
+  if [ "$LATTE_PERFORMANCE" = "yes" ]
+  then
+	test=`echo "($relativeTime-${performanceExpectedTimes[$name]})/${performanceExpectedTimes[$name]} < 0.1" | bc -l`
+	[ "$test" -eq 0 ] && exit -1
+  fi
 
   rm $REF energy.out input_tmp.in
